+++ weed/storage/backend/udm/api/private/v1/meta.proto
@@ -5,7 +5,7 @@ package private.v1;
 option go_package = "gitee.com/hhy-data/archive-api/private/v1;v1";
 
 
-enum TransportType {
+enum LocationType {
   unknown = 0;
 
   volume = 1;
@@ -21,4 +21,8 @@ enum CheckSumAlgorithm {
 message Checksum {
   CheckSumAlgorithm alg = 1;
   string digest = 2;
+}
+
+message DownloadFileResponse {
+  bytes data = 1;
 }
